Major library rewrite
The whole vision and scope of the library was reimagined, based on
real-life feedback from users. As a result, the most exotic and less
useful functionality (like the environment loader and the direct
binding to structs) were removed and the subcommand API was rebuilt
from scratch to be more clear and consistent with the main parser.

Overall, the end result is a more focused and lean implementation